+++ lazarus/cli/cli.py
@@ -6,7 +6,9 @@ from lazarus.bully import elect_leader
 from lazarus.cli.sink import app as sink_app
 from lazarus.sidecar import HeartbeatsListener
 from lazarus.cli.filter import app as filter_app
+from lazarus.cli.joiner import app as joiner_app
 from lazarus.cli.dataset import app as dataset_app
+from lazarus.cli.download import app as download_app
 from lazarus.constants import DEFAULT_HEARTBEAT_PORT
 from lazarus.cli.transform import app as transform_app
 from lazarus.docker_utils import SystemContainer, list_containers_from_config
@@ -18,6 +20,8 @@ app = typer.Typer()
 app.add_typer(dataset_app, name="dataset")
 app.add_typer(filter_app, name="filter")
 app.add_typer(transform_app, name="transform")
+app.add_typer(joiner_app, name="joiner")
+app.add_typer(download_app, name="download")
 app.add_typer(sink_app, name="sink")