+++ lazarus/storage/local.py
@@ -2,7 +2,7 @@ import json
 from pathlib import Path
 from io import TextIOWrapper
 from typing import Dict, Optional
-from collections import defaultdict
+from contextlib import contextmanager
 
 from lazarus.utils import get_logger
 from lazarus.exceptions import BadChecksumError
@@ -44,15 +44,35 @@ class LocalStorage(BaseStorage):
     def __init__(self, data_folder: Path):
         self.data_dir = data_folder
         self.data_dir.mkdir(parents=True, exist_ok=True)
-        self.data: Dict[TopicType, Dict[KeyType, MessageType]] = defaultdict(dict)
-        self.data_files: Dict[TopicType, TextIOWrapper] = {}
+        self.data: Dict[TopicType, Dict[KeyType, MessageType]] = dict()
+        self.data_files: Dict[TopicType, TextIOWrapper] = dict()
+        self._recovery_mode = False  # TODO: move this to the base class
+
+    @property
+    def in_recovery_mode(self):
+        return self._recovery_mode
+
+    @contextmanager
+    def recovery_mode(self):
+        """Recovery mode ensures no changes to the storage will be made
+        while allowing to fetch data from it."""
+        # Move to base class
+        self._recovery_mode = True
+        yield self
+        self._recovery_mode = False
 
     def __del__(self):
         for v in self.data_files.values():
             v.close()
 
     def contains(self, key: KeyType, topic: Optional[TopicType] = None):
-        return key in self.data[topic]
+        try:
+            return key in self.data[topic]
+        except KeyError:
+            return False
+
+    def contains_topic(self, topic: TopicType):
+        return topic in self.data
 
     def put(
         self,
@@ -61,6 +81,9 @@ class LocalStorage(BaseStorage):
         topic: Optional[TopicType] = None,
         autosync: bool = True,
     ):
+        if self._recovery_mode:
+            return
+
         if topic not in self.data_files:
             self.data_files[topic] = open(self.data_dir / f"{topic}.jsonl", "w")
 
@@ -74,12 +97,24 @@ class LocalStorage(BaseStorage):
         )
         self.data_files[topic].write("\n")
 
+        if topic not in self.data:
+            self.data[topic] = {}
+
         self.data[topic][key] = message
 
         if autosync:
             self.sync(topic)
 
+    def iter_topic(self, topic: TopicType):
+        if topic not in self.data:
+            raise KeyError(f"Topic {topic} is not present")
+        for k, v in self.data[topic].items():
+            yield k, v
+
     def sync(self, topic: Optional[TopicType] = None):
+        if self._recovery_mode:
+            return
+
         if topic is not None:
             self.data_files[topic].flush()
         else: