+++ tests/test_storage.py
@@ -63,3 +63,33 @@ def test_load_from_disk_step_on_keys(storage):
     storage.put("k1", "v2", topic="topic")
     new_storage = LocalStorage.load(storage.data_dir)
     assert new_storage.get("k1", topic="topic") == "v2"
+
+
+def test_add_int(storage):
+    storage.put("k1", 1)
+    new_storage = LocalStorage.load(storage.data_dir)
+    assert storage.get("k1") == 1
+    assert new_storage.get("k1") == 1
+
+
+def test_add_dict(storage):
+    storage.put("k1", {"v1": 1, "v2": "2"})
+    new_storage = LocalStorage.load(storage.data_dir)
+    assert storage.get("k1") == {"v1": 1, "v2": "2"}
+    assert new_storage.get("k1") == {"v1": 1, "v2": "2"}
+
+
+def test_recovery_mode(storage):
+    storage.put("k", "v1", topic="topic")
+    assert storage.get("k", topic="topic") == "v1"
+    storage.put("k", "v2", topic="topic")
+    assert storage.get("k", topic="topic") == "v2"
+    with storage.recovery_mode():
+        everything = [(k, v) for k, v in storage.iter_topic("topic")]
+        assert everything == [("k", "v2")]
+        storage.put("k", "v3", topic="topic")
+        everything = [(k, v) for k, v in storage.iter_topic("topic")]
+        assert everything == [("k", "v2")]
+    assert storage.get("k", topic="topic") == "v2"
+    storage.put("k", "v4", topic="topic")
+    assert storage.get("k", topic="topic") == "v4"